fix for BB2-128, move error handling out of views.py into exceptions.py
