+++ Makefile
@@ -3,7 +3,7 @@ reqs-compile:
 	pip-compile --generate-hashes --output-file requirements/requirements.dev.txt requirements/requirements.dev.in
 
 reqs-download:
-	pip download -r requirements/requirements.dev.txt --dest vendor --platform linux_x86_64 --no-deps
+	pip download -r requirements/requirements.txt --dest vendor --platform linux_x86_64 --no-deps
 
 reqs-install:
 	pip install -r requirements/requirements.txt --no-index --find-links ./vendor/
